turned reviews into a functional component and added props
